+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Version 1.0
 
+- (B.7) Game mat: Added tests for magic card placement and refactored common code
 - (B.6) Game mat: Implemented monster zone max capacity
 - (B.5) Game mat: Monster zone max cards failing test 
 - (B.4) Game mat: Monster zone passing test (wrong implementation)